requirements.txt
Django==1.10.5
django-cors-headers==2.0.0
djangorestframework==3.5.3
gunicorn==19.6.0
